Change the result's style behavior
Now it'll be fully white, then become greyed out after the first use of
the function, either by clicking "find gcd" or "clear".
This is because the text matters only at the start, after that it's not
as important, as the user will know the result will be displayed there.